test: consolidate subject tests into existing files
Move TestSearchBySANSubject into e2e_test.go and the SearchIndexHandler
subject tests into index_test.go per review feedback; drop the separate
*_subject_test.go files.
